+++ turbo/apps/cli/CHANGELOG.md
@@ -1,5 +1,16 @@
 # Changelog
 
+## [9.145.4](https://github.com/vm0-ai/vm0/compare/cli-v9.145.3...cli-v9.145.4) (2026-05-08)
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * devDependencies
+    * @vm0/api-contracts bumped to 1.1.0
+    * @vm0/connectors bumped to 1.1.0
+    * @vm0/core bumped to 8.262.0
+
 ## [9.145.3](https://github.com/vm0-ai/vm0/compare/cli-v9.145.2...cli-v9.145.3) (2026-05-08)